move buildValidationSchema to the hook
